+++ src/main/java/com/ss/academy/java/service/book/BookHistoryService.java
@@ -19,9 +19,9 @@ public interface BookHistoryService {
 	
 	List<BookHistory> findAllBooksHistory(Integer offset, Integer maxResults);
 	
-	List<BookHistory> findAllBooksHistory(Integer offset, Integer maxResults, Integer isReturned);
+	List<BookHistory> findAllBooksHistory(Integer offset, Integer maxResults, byte isReturned);
 	
 	Long countAllBooksHistory();
 	
-	Long countAllBooksHistory(Integer isReturned);
+	Long countAllBooksHistory(byte isReturned);
 }
